+++ src/com/nika/salad/vegetable/nightshade/Nightshade.java
@@ -7,7 +7,7 @@ import com.nika.salad.vegetable.Vitamins;
  * Created by nika_shkuratava on 3/6/2017.
  */
 public class Nightshade extends Vegetable {
-    public Nightshade(int calories, double carbohydrates, double proteins, Vitamins[] vitamins) {
+    public Nightshade(double calories, double carbohydrates, double proteins, Vitamins[] vitamins) {
         super(calories, carbohydrates, proteins, vitamins);
     }
 }
